pyproject.toml
[project]
name = "async-collab"
version = "0.1.0"
description = "PeopleJoin: LM Agents for Coordinating Multi-User Information Gathering"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # async-friendly version of flask, for the back-end API
    "quart>=0.19.4",
    # for deploying the API
    "hypercorn>=0.15.0",
    "jsons>=1.6.3",
    "pint>=0.23",
    "openai>=1.13.3",
    "websockets>=12.0",
    "msgraph-sdk>=1.2.0",
    "pythonnet>=3.0.3",
    "python-dotenv>=1.0.1",
    "websocket-client>=1.8.0",
    "pytz>=2024.1",
    "types-pytz>=2024.1.0.20240417",
    "rank_bm25>=0.2.2",
    # following are from sps
    "find-libpython>=0.3.0",
    "pyrsistent>=0.19.3",
    "msal>=1.28.0",
    "pyyaml>=6.0.1",
    "datasets>=2.14.5",
    "jsonnet-binary==0.17.0",
    "typer>=0.9.0",
    "cattrs>=23.1.2",
    "streamlit>=1.27.1",
    "openpyxl>=3.1.2",
    "nest-asyncio>=1.5.8",
    "evaluate>=0.4.1",
    "cer>=1.2.0",
    "streamlit-tree-select>=0.0.5",
    "azure-identity>=1.15.0",
    "azure-keyvault-secrets>=4.7.0",
    "sqlparse>=0.5.1",
    "matplotlib>=3.9.2",
    "aiohttp>=3.0",
    "platformdirs>=3.0",
    "tiktoken>=0.7.0",
    "lark>=1.1",
    "more-itertools>=8",
    "numpy>=1.21",
    "regex>=2023.6.3",
    "scipy>=1.11.1",
    "python-pptx>=1.0.2",
    "nltk>=3.9.1",
]

[project.optional-dependencies]
dev = [
    "black>=23.3.0",
    "pyright>=1.1.303",
    "pytest>=7.3.1",
    "ruff>=0.0.261",
    "pytest-asyncio>=0.21.1",
]



[tool.black]
skip-magic-trailing-comma = true
target-version = ["py311"]

[tool.pytest.ini_options]
asyncio_mode = "auto"



[tool.pyright]
include = [
  "src",
  "tests",
]
reportUnnecessaryCast = "error"
reportUnnecessaryTypeIgnoreComment = "error"

[tool.ruff]
# See hhttps://beta.ruff.rs/docs/rules/ for a list of rules.
# This list is kept in the same order as the documentation.
select = [
  "E",
  "F",
  "W",
  "I",
  "UP",
  "B",
  "C4",
  "RUF",
]
ignore = [
  # Do not perform function call in argument defaults	
  "B008",
  # Line too long
  "E501",
]
target-version = "py311"
src = [
  "src",
  "tests",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
